patch for windows warnings
git-svn-id: svn://svn.savannah.nongnu.org/osip/trunk/osip@625 a617929b-4c5a-41d8-b177-2eda7db2ca6c

+++ src/osip2/osip.c
@@ -1779,7 +1779,7 @@ osip_timers_gettimeout (osip_t * osip, struct timeval *lower_tv)
      {
        struct timeval cmpTime;
        div_t dValue = div(ixt->interval, 1000);
-       cmpTime.tv_sec = ixt->start + dValue.quot;
+       cmpTime.tv_sec = (long)(ixt->start + dValue.quot);
        cmpTime.tv_usec = dValue.rem * 1000;
        min_timercmp(lower_tv, &cmpTime);
