+++ src/osip2/osip_time.c
@@ -62,7 +62,7 @@ osip_gettimeofday (struct timeval *tp, void *tz)
   struct _timeb timebuffer;
 
   _ftime (&timebuffer);
-  tp->tv_sec = timebuffer.time;
+  tp->tv_sec = (long)timebuffer.time;
   tp->tv_usec = timebuffer.millitm * 1000;
   return 0;
 }
